move coursecopeservice into services folder
